+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `SVGCursor` can now be scaled to increase or decreased the size of the cursor.
 - New API to reset the interaction settings to their defaults.
+- `SVGCursor.SetCursorOptimise(bool)` allows control over whether the SVG cursor is rendered optimised for speed or not.
 
 ## [1.4.0] - 2023-04-06